Add AP test
