+++ src/remote/activitypub/resolver.ts
@@ -69,40 +69,3 @@ export default class Resolver {
 		return object;
 	}
 }
-
-type MockResponse = {
-	type: string;
-	content: string;
-};
-
-export class MockResolver extends Resolver {
-	//#region For mock
-	private _rs = new Map<string, MockResponse>();
-	public async _register(uri: string, content: string | Object, type = 'application/activity+json') {
-		this._rs.set(uri, {
-			type,
-			content: typeof content === 'string' ? content : JSON.stringify(content)
-		});
-	}
-	//#endregion
-
-	//#region Overrides
-	public async resolve(value: string | IObject): Promise<IObject> {
-		if (typeof value !== 'string') return value;
-
-		const r = this._rs.get(value);
-
-		if (!r) {
-			throw {
-				name: `StatusError`,
-				statusCode: 404,
-				message: `Not registed for mock`
-			};
-		}
-
-		const object = JSON.parse(r.content);
-
-		return object;
-	}
-	//#endregion
-}
